combine useGroupActions into single hook, break out navigation into two new hooks
